Make unassign button work
